Lifted out api call to context for better call
